+++ cxx/vineyard/filter.cpp
@@ -14,7 +14,7 @@ inline float Tukey(float const &r, float const &sw, float const &ew, float const
   if (r > ew) {
     return 0.f;
   } else if (r > sw) {
-    return (0.5f * ((1.f + eh) + (1.f - eh) * std::cosf((std::numbers::pi_v<float> * (r - sw)) / (ew - sw))));
+    return (0.5f * ((1.f + eh) + (1.f - eh) * std::cos((std::numbers::pi_v<float> * (r - sw)) / (ew - sw))));
   } else {
     return 1.f;
   }
